Retirado Navigation Drawer e adicionado activities. Lista de repositórios e pull requests funcionando.
