--- app/src/main/java/br/com/victoramaral/githubdive/view/activities/MainActivity.java
@@ -1,68 +1,0 @@
-package br.com.victoramaral.githubdive.view.activities;
-
-import android.os.Bundle;
-import android.view.View;
-import android.widget.TextView;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
-import androidx.navigation.ui.AppBarConfiguration;
-import androidx.navigation.ui.NavigationUI;
-
-import com.google.android.material.navigation.NavigationView;
-
-import br.com.victoramaral.githubdive.R;
-
-import static br.com.victoramaral.githubdive.view.activities.LoginActivity.EMAIL_KEY;
-import static br.com.victoramaral.githubdive.view.activities.LoginActivity.NOME_KEY;
-
-public class MainActivity extends AppCompatActivity {
-
-    private AppBarConfiguration mAppBarConfiguration;
-    private TextView textNomeUsuario;
-    private TextView textEmailUsuario;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        NavigationView navigationView = findViewById(R.id.nav_view);
-        View headerView = navigationView.getHeaderView(0);
-
-        mAppBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.nav_principal, R.id.nav_sobre)
-                .setDrawerLayout(drawer)
-                .build();
-
-        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
-        NavigationUI.setupActionBarWithNavController(this, navController, mAppBarConfiguration);
-        NavigationUI.setupWithNavController(navigationView, navController);
-
-        textNomeUsuario = headerView.findViewById(R.id.text_nome);
-        textEmailUsuario = headerView.findViewById(R.id.text_email);
-
-        if (getIntent() != null) {
-            Bundle bundle = getIntent().getExtras();
-            String nome = bundle.getString(NOME_KEY);
-            String email = bundle.getString(EMAIL_KEY);
-            textNomeUsuario.setText(nome);
-            textEmailUsuario.setText(email);
-        }
-
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
-        return NavigationUI.navigateUp(navController, mAppBarConfiguration)
-                || super.onSupportNavigateUp();
-    }
-}
