+++ app/src/main/java/br/com/victoramaral/githubdive/view/activities/SplashActivity.java
@@ -30,7 +30,7 @@ public class SplashActivity extends AppCompatActivity {
     private void jump() {
         timer.cancel();
         startActivity(new Intent(SplashActivity.this,
-                LoginActivity.class));
+                RepositoriesActivity.class));
         finish();
     }
 }
