--- app/src/main/java/br/com/victoramaral/githubdive/viewmodel/SobreViewModel.java
@@ -1,19 +1,0 @@
-package br.com.victoramaral.githubdive.viewmodel;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class SobreViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public SobreViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is send fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
