+++ app/src/main/res/layout/item_recycler_view_request.xml
@@ -19,7 +19,7 @@
             android:layout_marginTop="12dp"
             android:text="Título pull request"
             android:textColor="@color/textTitulo"
-            android:textSize="18sp"
+            android:textSize="20sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.085"
@@ -28,14 +28,15 @@
 
         <TextView
             android:id="@+id/textViewDescricao"
-            android:layout_width="380dp"
-            android:layout_height="60dp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
             android:layout_marginStart="10dp"
             android:layout_marginLeft="10dp"
             android:layout_marginTop="4dp"
             android:ellipsize="end"
-            android:maxLines="3"
-            android:text="body do pull request chamada api githubrepositório chamada api github"
+            android:maxLines="2"
+            android:text="body do pull request chamada api githubrepositório chamada api githubbody do pull request chamada api githubrepositório chamada api githubbody do pull request chamada api githubrepositório chamada api github"
+            android:textSize="16sp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.151"
             app:layout_constraintStart_toStartOf="parent"
@@ -43,32 +44,32 @@
 
         <de.hdodenhof.circleimageview.CircleImageView
             android:id="@+id/imageViewProfile"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_marginTop="8dp"
+            android:layout_width="65dp"
+            android:layout_height="65dp"
             android:src="@drawable/githublogobranco"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="@+id/textViewDescricao"
             app:layout_constraintTop_toBottomOf="@+id/textViewDescricao" />
 
         <TextView
             android:id="@+id/textViewUsername"
             android:layout_width="wrap_content"
-            android:layout_height="17dp"
-            android:layout_marginStart="42dp"
+            android:layout_height="wrap_content"
             android:layout_marginLeft="42dp"
-            android:layout_marginTop="16dp"
             android:text="Username"
             android:textColor="@color/textUserName"
-            android:textSize="14sp"
+            android:textSize="16sp"
+            app:layout_constraintBottom_toBottomOf="@+id/imageViewProfile"
             app:layout_constraintStart_toEndOf="@+id/imageViewProfile"
-            app:layout_constraintTop_toBottomOf="@+id/textViewDescricao" />
+            app:layout_constraintTop_toTopOf="@+id/imageViewProfile"
+            app:layout_constraintVertical_bias="0.17000002" />
 
         <TextView
             android:id="@+id/textViewNomeSobrenome"
             android:layout_width="wrap_content"
-            android:layout_height="16dp"
+            android:layout_height="wrap_content"
             android:text="Nome e sobrenome"
-            android:textSize="12sp"
+            android:textSize="14sp"
             app:layout_constraintEnd_toEndOf="@+id/textViewUsername"
             app:layout_constraintStart_toStartOf="@+id/textViewUsername"
             app:layout_constraintTop_toBottomOf="@+id/textViewUsername" />
