--- app/src/main/res/menu/activity_main_drawer.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/nav_principal"
-            android:icon="@drawable/ic_menu_principal"
-            android:title="@string/menu_principal" />
-    </group>
-
-    <item android:title="Informações">
-        <menu>
-            <item
-                android:id="@+id/nav_sobre"
-                android:icon="@drawable/ic_menu_sobre"
-                android:title="@string/menu_sobre" />
-        </menu>
-    </item>
-
-</menu>
