--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_principal">
-
-    <fragment
-        android:id="@+id/nav_principal"
-        android:name="br.com.victoramaral.githubdive.view.fragments.PrincipalFragment"
-        android:label="@string/menu_principal"
-        tools:layout="@layout/fragment_principal" />
-
-    <fragment
-        android:id="@+id/nav_sobre"
-        android:name="br.com.victoramaral.githubdive.view.fragments.SobreFragment"
-        android:label="@string/menu_sobre"
-        tools:layout="@layout/fragment_sobre" />
-</navigation>
