+++ app/src/main/res/values/strings.xml
@@ -1,13 +1,3 @@
 <resources>
     <string name="app_name">GitHub Dive</string>
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">Usuário</string>
-    <string name="nav_header_subtitle">Email</string>
-    <string name="nav_header_desc">Navigation header</string>
-    <string name="action_sair">Sair do aplicativo</string>
-
-    <string name="menu_principal">Principal</string>
-    <string name="menu_sobre">Sobre</string>
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
